+++ internal/dialectadapter/dialectquery/dialectquery.go
