+++ internal/dialectadapter/dialectquery/redshift.go
