+++ internal/dialectadapter/dialectquery/sqlite3.go
