+++ internal/dialectadapter/dialectquery/vertica.go
